[Bronze V] Title: Multiply, Time: 168 ms, Memory: 17628 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -6,8 +6,8 @@ public class Main {
         
         int a = sc.nextInt();
         int b = sc.nextInt();
-        int c = sc.nextInt();
-        int d = sc.nextInt();
+        long c = sc.nextLong();
+        long d = sc.nextLong();
         
         System.out.println(c * d);
